Create dropCollection test utils
